add reuseable onToggleColumnsColorScales method
